--- config/conceptnet/eval_changes.json
@@ -1,17 +1,0 @@
-{
-
-    "base": {
-        "0": {
-            "gpu_index": 0
-        },
-        "1": {
-            "gpu_index": 1
-        },
-        "2": {
-            "gpu_index": 2
-        },
-        "3": {
-            "gpu_index": 3
-        }
-    }
-}
