--- scripts/setup/get_atomic_data.sh
@@ -1,6 +1,0 @@
-wget https://homes.cs.washington.edu/~msap/atomic/data/atomic_data.tgz
-mkdir -p data/atomic
-mv atomic_data.tgz data/atomic
-
-tar -xvzf data/atomic/atomic_data.tgz -C data/atomic
-rm data/atomic/atomic_data.tgz
