--- scripts/setup/get_model_files.sh
@@ -1,3 +1,0 @@
-git clone https://github.com/openai/finetune-transformer-lm.git
-mv finetune-transformer-lm/model .
-rm -rf finetune-transformer-lm
