+++ src/config.py
@@ -1,5 +1,5 @@
 import json
-from utils.utils import DD
+import copy
 
 device = "cpu"
 
@@ -184,3 +184,34 @@ def load_config(name):
     with open(name, "r") as f:
         config = json.load(f)
     return config
+
+
+# Taken from Jobman 0.1
+class DD(dict):
+    def __getattr__(self, attr):
+        if attr == '__getstate__':
+            return super(DD, self).__getstate__
+        elif attr == '__setstate__':
+            return super(DD, self).__setstate__
+        elif attr == '__slots__':
+            return super(DD, self).__slots__
+        return self[attr]
+
+    def __setattr__(self, attr, value):
+        # Safety check to ensure consistent behavior with __getattr__.
+        assert attr not in ('__getstate__', '__setstate__', '__slots__')
+#         if attr.startswith('__'):
+#             return super(DD, self).__setattr__(attr, value)
+        self[attr] = value
+
+    def __str__(self):
+        return 'DD%s' % dict(self)
+
+    def __repr__(self):
+        return str(self)
+
+    def __deepcopy__(self, memo):
+        z = DD()
+        for k, kv in self.items():
+            z[k] = copy.deepcopy(kv, memo)
+        return z
